# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/guides/{guideId}/votes/summary": {
      "get": {
        "operationId": "getGuideVoteSummary",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideVoteSummaryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Guide was not found."
          }
        },
        "summary": "Get guide vote summary",
        "tags": [
          "community"
        ]
      }
    },
    "/guides/{guideId}/vote": {
      "put": {
        "operationId": "upsertGuideVote",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertGuideVoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideVoteSummaryResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          },
          "404": {
            "description": "Guide was not found."
          }
        },
        "summary": "Create or update current user guide vote",
        "tags": [
          "community"
        ]
      },
      "delete": {
        "operationId": "removeGuideVote",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideVoteSummaryResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          },
          "404": {
            "description": "Guide was not found."
          }
        },
        "summary": "Remove current user guide vote",
        "tags": [
          "community"
        ]
      }
    },
    "/guides/{guideId}/comments": {
      "get": {
        "operationId": "listGuideComments",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentListResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Guide was not found."
          }
        },
        "summary": "List visible guide comments",
        "tags": [
          "community"
        ]
      },
      "post": {
        "operationId": "createGuideComment",
        "parameters": [
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          }
        },
        "summary": "Create a guide comment",
        "tags": [
          "community"
        ]
      }
    },
    "/guides/{guideId}/comments/{commentId}": {
      "patch": {
        "operationId": "updateGuideComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "User cannot manage this comment."
          }
        },
        "summary": "Update a guide comment",
        "tags": [
          "community"
        ]
      },
      "delete": {
        "operationId": "deleteGuideComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "guideId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "User cannot manage this comment."
          }
        },
        "summary": "Soft-delete a guide comment",
        "tags": [
          "community"
        ]
      }
    },
    "/sessions/{sessionId}/comments": {
      "get": {
        "operationId": "listSessionComments",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentListResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Session was not found or is private."
          }
        },
        "summary": "List visible session comments",
        "tags": [
          "community"
        ]
      },
      "post": {
        "operationId": "createSessionComment",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          }
        },
        "summary": "Create a session comment",
        "tags": [
          "community"
        ]
      }
    },
    "/sessions/{sessionId}/comments/{commentId}": {
      "patch": {
        "operationId": "updateSessionComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "User cannot manage this comment."
          }
        },
        "summary": "Update a session comment",
        "tags": [
          "community"
        ]
      },
      "delete": {
        "operationId": "deleteSessionComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "User cannot manage this comment."
          }
        },
        "summary": "Soft-delete a session comment",
        "tags": [
          "community"
        ]
      }
    },
    "/reports": {
      "post": {
        "operationId": "createContentReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentReportResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No active session."
          },
          "404": {
            "description": "Reported content was not found."
          }
        },
        "summary": "Create a moderation intake report",
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Steam Achievement Tracker API",
    "description": "API for syncing and reading Steam achievement intelligence data",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Docker backend"
    }
  ],
  "components": {
    "schemas": {
...
      "GuideVoteSummaryResponseDto": {
        "type": "object",
        "properties": {
          "upvotes": {
            "type": "number",
            "example": 12
          },
          "downvotes": {
            "type": "number",
            "example": 2
          },
          "score": {
            "type": "number",
            "example": 10
          },
          "currentUserVote": {
            "type": "number",
            "enum": [
              -1,
              1
            ],
            "nullable": true
          }
        },
        "required": [
          "upvotes",
          "downvotes",
          "score"
        ]
      },
      "UpsertGuideVoteDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "enum": [
              -1,
              1
            ],
            "example": 1
          }
        },
        "required": [
          "value"
        ]
      },
      "CommunityAuthorResponseDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "steamId": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "publicSlug": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "example": "visible"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/CommunityAuthorResponseDto"
          }
        },
        "required": [
          "id",
          "body",
          "status",
          "createdAt",
          "updatedAt",
          "author"
        ]
      },
      "CommentListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponseDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "body"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "body"
        ]
      },
      "CreateContentReportDto": {
        "type": "object",
        "properties": {
          "targetType": {
            "type": "string",
            "enum": [
              "guide",
              "guide_comment",
              "gaming_session",
              "session_comment"
            ]
          },
          "targetId": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "enum": [
              "spam",
              "abuse",
              "off_topic",
              "cheating",
              "other"
            ]
          },
          "details": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          }
        },
        "required": [
          "targetType",
          "targetId",
          "reason"
        ]
      },
      "ContentReportResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "targetType": {
            "type": "string",
            "enum": [
              "guide",
              "guide_comment",
              "gaming_session",
              "session_comment"
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "spam",
              "abuse",
              "off_topic",
              "cheating",
              "other"
            ]
          },
          "status": {
            "type": "string",
            "example": "open"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "targetType",
          "reason",
          "status",
          "createdAt"
        ]
      },
...
    }
  }
}